fix: Use pubdate for video publication time (BiliBili)
The extractor was previously using the `ctime` field for determining a video's publication date. However, analysis of the Bilibili API shows that `ctime` represents the creation time of the data entry and can be unreliable, often showing June 2017 for older videos.

This commit changes the implementation to use the `pubdate` field instead.

See: https://github.com/SocialSisterYi/bilibili-API-collect/issues/306